Break out iterator XML tree parsers and improve naming, among other tweaks.
