+++ pylib/uxml/version.py
@@ -1,2 +1,2 @@
-#http://legacy.python.org/dev/peps/pep-0440/
-version_info = ('3', '0', '2')
+# http://legacy.python.org/dev/peps/pep-0440/
+version_info = ('3', '1', '0')
