+++ pylib/uxml/xml.py
@@ -1,20 +1,25 @@
+# -----------------------------------------------------------------------------
 # amara3.uxml.xml
+#
+# MicroXML tree objects parsed from XML sources
+#
+# -----------------------------------------------------------------------------
 
-#import asyncio
-from asyncio import coroutine
+import asyncio
 import xml.parsers.expat
-from xml.sax.saxutils import escape #also quoteattr?
+from xml.sax.saxutils import escape  # also quoteattr?
 
-from amara3.uxml import tree
-from amara3.uxml.parser import parse, parser, parsefrags, event
+from . import tree
+from .parser import parser, parsefrags, event
 
 
 class expat_callbacks(object):
-    def __init__(self, handler, asyncio_based_handler=True):
+    def __init__(self, handler, prime_handler=True):
         self._handler = handler
         self._elem_stack = []
-        if asyncio_based_handler:
-            next(self._handler) #Start the coroutine running
+        #if asyncio.iscoroutine(handler):
+        if prime_handler:
+            next(handler)  # Prime coroutine
         return
 
     def start_element(self, name, attrs):
@@ -102,15 +107,7 @@ def parse(source, handler):
     return p
 
 
-@coroutine
-def buffer_handler(accumulator):
-    while True:
-        event = yield
-        accumulator.append(event)
-    return
-
-
-#Tree-based tools
+# Tree-based tools
 
 class treebuilder(tree.treebuilder):
     '''
@@ -132,74 +129,3 @@ class treebuilder(tree.treebuilder):
         self.expat_parser.Parse(source)
         return self._root
 
-
-'''
-from asyncio import coroutine
-from amara3.uxml import xml
-@coroutine
-def sink(accumulator):
-    while True:
-        e = yield
-        accumulator.append(e.xml_value)
-
-values = []
-ts = xml.treesequence(('a', 'b'), sink(values))
-ts.parse('<a xmlns="urn:namespaces:suck"><b>1</b><b>2</b><b>3</b></a>')
-values
-
-----
-
-from asyncio import coroutine
-from amara3.uxml import tree
-from amara3.uxml.treeutil import *
-
-def ppath(start, path):
-    print((start, path))
-    if not path: return None
-    if len(path) == 1:
-        yield from select_name(start, path[0])
-    else:
-        for e in select_name(start, path[0]):
-            yield from ppath(e, path[1:])
-
-ts = tree.treebuilder()
-root = ts.parse('<a xmlns="urn:namespaces:suck"><b><c>1</c></b><b>2</b><b>3</b></a>')
-pathresults = ppath(root, ('b', 'c'))
-print(list(pathresults))
-'''
-
-class treesequence(tree.treesequence):
-    '''
-    >>> from asyncio import coroutine
-    >>> from amara3.uxml import xml
-    >>> @coroutine
-    ... def sink(accumulator):
-    ...     while True:
-    ...         e = yield
-    ...         accumulator.append(e.xml_value)
-    ...
-    >>> values = []
-    >>> ts = xml.treesequence(('a', 'b'), sink(values))
-    >>> ts.parse('<a xmlns="urn:namespaces:suck"><b>1</b><b>2</b><b>3</b></a>')
-    >>> values
-    ['1', '2', '3']
-    '''
-    def __init__(self, pattern, sink, callbacks=expat_callbacks):
-        super(treesequence, self).__init__(pattern, sink)
-        self.handler = callbacks(self._handler())
-        self.expat_parser = xml.parsers.expat.ParserCreate(namespace_separator=' ')
-
-        self.expat_parser.StartElementHandler = self.handler.start_element
-        self.expat_parser.EndElementHandler = self.handler.end_element
-        self.expat_parser.CharacterDataHandler = self.handler.char_data
-        self.expat_parser.StartNamespaceDeclHandler = self.handler.start_namespace
-        self.expat_parser.EndNamespaceDeclHandler = self.handler.end_namespace
-        return
-
-    def parse(self, source):
-        self.expat_parser.Parse(source)
-        return
-
-    def parse_file(self, fp):
-        self.expat_parser.ParseFile(fp)
-        return
